Removed C++ standard requirement from Makevars. See: https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Using-C_002b_002b-code
"On a Unix-alike C++ compilation defaulted to C++11 from R 3.6.0, to C++14 from R 4.1.0 and to C++17 from R 4.3.0. However, only ?if available?, so platforms using very old OSes might have used the previous default. Even older versions of R defaulted to the compiler?s default, almost certainly C++98 for compilers of comparable vintage.
On Windows the default was changed from C++98 to C++11 in R 3.6.2, to C++14 in R 4.2.3 and to C++17 in R 4.3.0. "

This is targeting R >= 4.1.0 so the default C++ standard is fine. The minimum
required by the nanoflann.hpp library is C++11.